Reapply "refactor"
This reverts commit afd91d94175a6d1398a6e002bafbc40cf25d087f.
