+++ code/algorithm-demo/src/com/noob/algorithm/base/template/Main01.java
@@ -4,7 +4,7 @@ package com.noob.algorithm.base.template;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Main {
+public class Main01 {
 
     // 初始化测试用例
     public static List<int[]> initCase() {
